backend/requirements.txt
# AI Mustaqbil 2.0 — Backend Python Dependencies
# Install with: pip install -r requirements.txt

# ── Server ───────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12    # FastAPI file uploads

# ── LangGraph / LangChain ────────────────────────────────────
langgraph==0.2.50
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
langchain-openai==0.2.6     # OpenRouter uses OpenAI-compatible API

# ── LLM SDKs ─────────────────────────────────────────────────
openai==1.54.0              # Used for OpenRouter (DeepSeek-R1)
groq==0.11.0                # Used for Whisper + Llama-3-8B

# ── Vector Store / RAG ───────────────────────────────────────
faiss-cpu==1.9.0
chromadb==0.5.20
pypdf==4.3.1                # PDF ingestion for WHO guidelines
tiktoken==0.8.0             # Token counting for chunking

# ── Schema & Config ──────────────────────────────────────────
pydantic==2.9.2
python-dotenv==1.0.1

# ── Utilities ────────────────────────────────────────────────
httpx==0.27.2
aiofiles==24.1.0
